requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pydantic>=2.0
openai>=1.0.0